Update chip to trigger popover instead of button
